frons-mcp-server/requirements.txt
fastmcp>=2.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
supabase>=2.13.0
python-dotenv>=1.0.0
